FILE: moves validators folder into config-form folder
