+++ libs/spirit-islander/config/feature/src/lib/config-form/config-form/config-form.ts
@@ -17,7 +17,7 @@ import {
   playersOutnumberTotalBoards,
   required,
   restrictedBoardPairings,
-} from '../validators';
+} from './validators';
 import {
   Settings,
   createDefaultSettings,
