+++ libs/spirit-islander/config/feature/src/lib/config-form/config-form/validators/invalid-difficulty-range.ts
