+++ libs/spirit-islander/config/feature/src/lib/config-form/config-form/validators/players-outnumber-spirits.ts
