+++ libs/spirit-islander/config/feature/src/lib/config-form/config-form/validators/restricted-board-pairings.spec.ts
